requirements.txt
adbc_driver_flightsql==1.3.0
adbc-driver-manager==1.3.0
aiofiles==24.1.0
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
backoff==2.2.1
bcrypt==4.2.0
beautifulsoup4==4.13.4
blinker==1.9.0
boto3==1.39.7
botocore==1.39.7
build==1.2.2.post1
cachetools==5.5.0
certifi==2025.6.15
charset-normalizer==3.4.2
chroma-hnswlib==0.7.3
chromadb==1.1.0
click==8.1.7
coloredlogs==15.0.1
contourpy==1.3.0
cycler==0.12.1
dataclasses-json==0.6.7
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
deprecation==2.1.0
distro==1.9.0
docstring_parser==0.16
durationpy==0.9
Events==0.5
exceptiongroup==1.3.0
fastapi-offline==1.7.3
fastapi==0.115.12
fastmcp==2.12.4
filelock==3.16.1
Flask-HTTPAuth==4.8.0
Flask-Login==0.6.3
Flask==3.1.2
flatbuffers==24.3.25
fonttools==4.54.1
frozenlist==1.5.0
fsspec==2024.10.0
google-api-core==2.25.1
google-auth==2.35.0
google-cloud-aiplatform==1.100.0
google-cloud-bigquery==3.34.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.1.1
groq==0.11.0
grpc-google-iam-v1==0.14.2
grpcio==1.67.1
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx-sse==0.4.0
httpx==0.28.1
huggingface-hub==0.26.2
humanfriendly==10.0
idna==3.10
importlib_metadata==8.4.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.7.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.7
kubernetes==31.0.0
langchain-anthropic==0.3.10
langchain-aws==0.2.34
langchain-chroma==0.2.6
langchain-community==0.3.27
langchain-core==0.3.76
langchain-experimental==0.3.2
langchain-google-genai==2.0.9
langchain-google-vertexai==2.0.27
langchain-groq==0.2.0
langchain-mistralai==0.2.11
langchain-nvidia-ai-endpoints==0.3.3
langchain-ollama==0.3.6
langchain-openai==0.3.3
langchain-postgres==0.0.15
langchain-sambanova==0.1.2
langchain-text-splitters==0.3.8
langchain==0.3.26
langfuse==2.60.4
langsmith==0.4.4
lxml==6.0.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
matplotlib==3.9.2
mdurl==0.1.2
mmh3==5.0.1
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
numpy===1.26.4
oauthlib==3.2.2
ollama==0.5.3
onnxruntime==1.20.0; platform_system=="Windows"
openai==1.60.2
opensearch-py==2.7.1
opentelemetry-api==1.34.1
opentelemetry-exporter-otlp-proto-common==1.34.1
opentelemetry-exporter-otlp-proto-grpc==1.34.1
opentelemetry-instrumentation-asgi==0.55b1
opentelemetry-instrumentation-fastapi==0.55b1
opentelemetry-instrumentation==0.55b1
opentelemetry-proto==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
opentelemetry-util-http==0.55b1
orjson==3.10.18
overrides==7.7.0
packaging==24.2
pandas==2.2.2
pgvector==0.2.5
pillow==10.4.0
playwright==1.53.0
posthog==3.7.0
propcache==0.2.0
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
psycopg-pool==3.2.3
psycopg[binary,pool]==3.2.3
psycopg==3.2.3
pyarrow==19.0.1
pyasn1_modules==0.4.1
pyasn1==0.6.1
pydantic_core==2.33.2
pydantic-settings==2.6.1
pydantic==2.11.7
Pygments==2.18.0
pylance==0.37.0
pyparsing==3.2.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pysqlite3-binary==0.5.4; platform_system=="Linux"
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2024.1
PyYAML==6.0.2
ratelimiter==1.2.0.post0
regex==2024.9.11
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
requests==2.32.4
rich==13.9.4
rsa==4.9
s3transfer==0.13.0
safetensors==0.4.5
shapely==2.0.6
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sseclient-py==1.8.0
starlette==0.40.0
sympy==1.13.3
tenacity==9.1.2
tiktoken==0.7.0
tokenizers==0.21
tomli==2.0.2
tqdm==4.66.6
transformers==4.50.0
typer==0.15.2
types-requests
typing_extensions==4.14.1
typing-inspect==0.9.0
tzdata==2024.2
urllib3==2.5.0
uvicorn==0.32.0
waitress==3.0.2
watchfiles==0.24.0
websocket-client==1.8.0
websockets==14.0
Werkzeug==3.1.1
wrapt==1.16.0
yarl==1.17.1
zipp==3.20.2